// Singer.API/ClientApp/src/app/modules/core/models/singer-event-location.dto.ts
export class SingerEventLocation {
   id: string;
   name: string;
   address: string;
   postalCode: string;
   city: string;
   country: string;
}

export interface SingerEventLocationDTO {
   id: string;
   name: string;
   address: string;
   postalCode: string;
   city: string;
   country: string;
}

export interface CreateSingerEventLocationDTO {
   name: string;
   address: string;
   postalCode: string;
   city: string;
   country: string;
}

export interface UpdateSingerEventLocationDTO {
   name: string;
   address: string;
   postalCode: string;
   city: string;
   country: string;
}